+++ cms/djangoapps/contentstore/tests/test_core_caching.py
@@ -1,4 +1,7 @@
-from cache_toolbox.core import get_cached_content, set_cached_content, del_cached_content
+"""
+Tests core caching facilities.
+"""
+from contentserver.caching import get_cached_content, set_cached_content, del_cached_content
 from opaque_keys.edx.locations import Location
 from django.test import TestCase
