+++ cms/djangoapps/contentstore/views/assets.py
@@ -10,7 +10,7 @@ from django.views.decorators.http import require_http_methods, require_POST
 from django.conf import settings
 
 from edxmako.shortcuts import render_to_response
-from cache_toolbox.core import del_cached_content
+from contentserver.caching import del_cached_content
 
 from contentstore.utils import reverse_course_url
 from xmodule.contentstore.django import contentstore
