+++ common/djangoapps/contentserver/__init__.py
@@ -1,0 +1,5 @@
+"""
+Serves course assets to end users.
+"""
+
+CONTENTSERVER_VERSION = 1
